main: refactor code
Move functions into a separate file.
